+++ dwds/lib/src/loaders/frontend_server_require.dart
@@ -16,7 +16,7 @@ class FrontendServerRequireStrategyProvider {
   final PackageUriMapper _packageUriMapper;
   final Future<Map<String, String>> Function() _digestsProvider;
   final String _basePath;
-  final Uri? _appEntrypoint;
+  final BuildSettings _buildSettings;
 
   late final RequireStrategy _requireStrategy = RequireStrategy(
     _configuration,
@@ -28,7 +28,7 @@ class FrontendServerRequireStrategyProvider {
     _serverPathForAppUri,
     _moduleInfoForProvider,
     _assetReader,
-    _appEntrypoint,
+    _buildSettings,
   );
 
   FrontendServerRequireStrategyProvider(
@@ -36,7 +36,7 @@ class FrontendServerRequireStrategyProvider {
     this._assetReader,
     this._packageUriMapper,
     this._digestsProvider,
-    this._appEntrypoint,
+    this._buildSettings,
   ) : _basePath = _assetReader.basePath;
 
   RequireStrategy get strategy => _requireStrategy;