+++ dwds/lib/src/loaders/legacy.dart
@@ -71,7 +71,7 @@ class LegacyStrategy extends LoadStrategy {
   /// Returns `null` if not a google3 app.
   final String? Function(String absolutePath) _g3RelativePath;
 
-  final Uri? _appEntrypoint;
+  final BuildSettings _buildSettings;
 
   LegacyStrategy(
     this.reloadConfiguration,
@@ -81,7 +81,7 @@ class LegacyStrategy extends LoadStrategy {
     this._serverPathForAppUri,
     this._moduleInfoForProvider,
     AssetReader assetReader,
-    this._appEntrypoint,
+    this._buildSettings,
     this._g3RelativePath,
     String? packageConfigPath,
   ) : super(assetReader, packageConfigPath: packageConfigPath);
@@ -128,7 +128,7 @@ class LegacyStrategy extends LoadStrategy {
   String? serverPathForAppUri(String appUri) => _serverPathForAppUri(appUri);
 
   @override
-  Uri? get appEntrypoint => _appEntrypoint;
+  BuildSettings get buildSettings => _buildSettings;
 
   @override
   String? g3RelativePath(String absolutePath) => _g3RelativePath(absolutePath);