+++ dwds/lib/src/services/chrome_proxy_service.dart
@@ -174,22 +174,30 @@ class ChromeProxyService implements VmServiceInterface {
   }
 
   Future<void> _updateCompilerDependencies(String entrypoint) async {
-    final metadataProvider =
-        globalToolConfiguration.loadStrategy.metadataProviderFor(entrypoint);
-    final moduleFormat = globalToolConfiguration.loadStrategy.moduleFormat;
+    final loadStrategy = globalToolConfiguration.loadStrategy;
+    final moduleFormat = loadStrategy.moduleFormat;
+    final canaryFeatures = loadStrategy.buildSettings.canaryFeatures;
+    final experiments = loadStrategy.buildSettings.experiments;
+
+    // TODO(annagrin): Read null safety setting from the build settings.
+    final metadataProvider = loadStrategy.metadataProviderFor(entrypoint);
     final soundNullSafety = await metadataProvider.soundNullSafety;
 
     _logger.info('Initializing expression compiler for $entrypoint '
         'with sound null safety: $soundNullSafety');
 
+    final compilerOptions = CompilerOptions(
+      moduleFormat: moduleFormat,
+      soundNullSafety: soundNullSafety,
+      canaryFeatures: canaryFeatures,
+      experiments: experiments,
+    );
+
     final compiler = _compiler;
     if (compiler != null) {
-      await compiler.initialize(
-        moduleFormat: moduleFormat,
-        soundNullSafety: soundNullSafety,
-      );
-      final dependencies = await globalToolConfiguration.loadStrategy
-          .moduleInfoForEntrypoint(entrypoint);
+      await compiler.initialize(compilerOptions);
+      final dependencies =
+          await loadStrategy.moduleInfoForEntrypoint(entrypoint);
       await captureElapsedTime(
         () async {
           final result = await compiler.updateDependencies(dependencies);