+++ dwds/CHANGELOG.md
@@ -1,7 +1,8 @@
-## 22.2.0-wip
+## 23.0.0-wip
+- Restructure `LoadStrategy` to provide build settings. - [#2270](https://github.com/dart-lang/webdev/pull/2270)
 
 ## 22.1.0
-- Update `package:vm_service` constraint to `^13.0.0`. - [#2235](https://github.com/dart-lang/webdev/pull/2265)
+- Update `package:vm_service` constraint to `^13.0.0`. - [#2265](https://github.com/dart-lang/webdev/pull/2265)
 
 ## 22.0.0